linux-gen: modular-framework: move schedule subsystem header to a public folder
Move schedule subsystem header to a public folder.
