+++ include/Makefile.am
@@ -72,7 +72,8 @@ subsystemspecincludedir= $(includedir)/subsystem/spec
 subsystemspecinclude_HEADERS = \
 		  subsystem/spec/buffer_subsystem.h \
 		  subsystem/spec/pool_subsystem.h \
-		  subsystem/spec/queue_subsystem.h
+		  subsystem/spec/queue_subsystem.h \
+		  subsystem/spec/schedule_subsystem.h
 
 odpapiabidefaultincludedir= $(includedir)/odp/arch/default/api/abi
 odpapiabidefaultinclude_HEADERS = \
