+++ platform/linux-generic/schedule/sp.c
@@ -18,7 +18,7 @@
 #include <odp_config_internal.h>
 #include <odp_ring_internal.h>
 #include <odp_timer_internal.h>
-#include <odp_schedule_subsystem.h>
+#include <subsystem/spec/schedule_subsystem.h>
 #include <odp_packet_io_internal.h>
 #include <odp_queue_internal.h>
